backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==1.4.23
pydantic==1.8.2
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
numpy==1.26.1
